# microservices/high-five-service/src/api/api-spec.yaml
openapi: 3.0.1
info:
  title: high-five-service
  description: Service contract for high-five-service with OpenAPI 3.0 specifications
  version: 1.0.0
servers:
  - url: 'http://localhost:8000/high-fives/api'
  - url: 'http://localhost:8011/api'
  - url: 'http://18.217.249.183/high-fives/api'
...
security:
  - bearerAuth: []
  - x-userinfo: []
paths:
...
  /v1/highfive/{highfiveId}/notes/{noteId}:
    put:
      tags:
        - High Fives
        - Notes
      parameters:
        - name: highfiveId
          in: path
          description: ID of High Five Type
          required: true
          schema:
            type: string
        - name: noteId
          in: path
          description: ID of the note to be updated
          required: true
          schema:
            type: string
      requestBody:
        description: Note with updates
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotesBody'
        required: true
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotesResponse'
        404:
          description: HighFive or Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      x-codegen-request-body-name: NotesBody
    delete:
      tags:
        - High Fives
        - Notes
      parameters:
        - name: highfiveId
          in: path
          description: ID of High Five Type
          required: true
          schema:
            type: string
        - name: noteId
          in: path
          description: ID of the note to be updated
          required: true
          schema:
            type: string
      responses:
        204:
          description: Note deleted on High Five
        404:
          description: HighFive or Note not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: All requests against Kong Gateway must include the `Authorization` header containing your JWT Token.
    x-userinfo:
      type: apiKey
      in: header
      name: X-userinfo
      description: All requests against a standalone microservice must include the `x-userinfo` header containing your encoded user ID.
  schemas:
    ExtendedErrorInfo:
      required:
        - message
      type: object
      properties:
        message:
          type: string
        type:
          type: string
        key:
          type: string
        property:
          type: string
    Error:
      required:
        - message
      type: object
      properties:
        extraInfo:
          type: array
          items:
            $ref: '#/components/schemas/ExtendedErrorInfo'
        message:
          type: string
...
    NotesBody:
      type: object
      properties:
        content:
          type: string
        _parentId:
          type: string
      example:
        content: Test Name for High Five Type
...
    NotesResponse:
      type: object
      properties:
        _id:
          type: string
        _createdOn:
          type: string
        _createdBy:
          type: string
        _modifiedOn:
          type: string
        _modifiedBy:
          type: string
        _deleted:
          type: boolean
        content:
          type: string
        _parentId:
          type: string
      example:
        _id: a3b2d54f-e3d5-49a5-9fde-107ce4e34c91
        _createdOn: 2021-05-10T16:36:06.326Z
        _createdBy: 61087fbe-958a-4e21-865e-910c57141a91
        _modifiedOn: 2021-05-09T15:36:06.326Z
        _modifiedBy: c6681ac5-224c-479e-994b-44dbc3cc3f36
        _deleted: false
        content: A sample Note
...